+++ intern/src/editor_gui/edit_inspector_transformation.cpp
@@ -7,7 +7,7 @@ namespace Edit
 {
     CInspectorTransformation::CInspectorTransformation(QWidget* _pParent)
         : QWidget          (_pParent)
-        , m_CurrentEntityID(-1)
+        , m_CurrentEntityID(static_cast<unsigned int>(-1))
     {
         // -----------------------------------------------------------------------------
         // Setup UI
